[fix] fix decode module

+++ src/cpu/decode/inst_32/a_extension.rs
@@ -2,7 +2,7 @@ use crate::cpu::decode::DecodeUtil;
 use crate::cpu::instruction::OpecodeKind;
 use crate::cpu::TrapCause;
 
-pub fn parse_opecode(inst: u64) -> Result<OpecodeKind, &'static str> {
+pub fn parse_opecode(inst: u32) -> Result<OpecodeKind, &'static str> {
     let opmap: u8 = inst.slice(6, 0) as u8;
     let funct7: u8 = inst.slice(31, 27) as u8;
 
@@ -26,7 +26,7 @@ pub fn parse_opecode(inst: u64) -> Result<OpecodeKind, &'static str> {
 }
 
 pub fn parse_rd(
-    inst: u64,
+    inst: u32,
     opkind: &OpecodeKind,
 ) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rd: usize = inst.slice(11, 7) as usize;
@@ -48,7 +48,7 @@ pub fn parse_rd(
 }
 
 pub fn parse_rs1(
-    inst: u64,
+    inst: u32,
     opkind: &OpecodeKind,
 ) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rs1: usize = inst.slice(19, 15) as usize;
@@ -70,7 +70,7 @@ pub fn parse_rs1(
 }
 
 pub fn parse_rs2(
-    inst: u64,
+    inst: u32,
     opkind: &OpecodeKind,
 ) -> Result<Option<usize>, (Option<u64>, TrapCause, String)> {
     let rs2: usize = inst.slice(24, 20) as usize;
@@ -92,7 +92,7 @@ pub fn parse_rs2(
 
 #[allow(non_snake_case)]
 pub fn parse_imm(
-    inst: u64,
+    inst: u32,
     opkind: &OpecodeKind,
 ) -> Result<Option<i32>, (Option<u64>, TrapCause, String)> {
     let aq_and_rl = || inst.slice(26, 25) as i32;
